Fix big-endian length handling in msgpack
Most of the code was not working properly when handling data size. The
reason is that msgpack implementation was handling data length as
big-endian only in certain occasions, while msgpack implementation
requires to threat data length _ALWAYS_ as big-endian. This patch should
fix this issue and introduce more cleaniness inside the utils.h header.