Handle SIGCHLD via signal()
Using signalfd() was pointless, since the syscall doesn't provide
support for multiple process and when we read signalfd() file
descriptor, we read() only the last completed process. This obviously
reflects on the code, which normally uses waitpid() to loop over all
process children. We switch to signal() usage, so we can handle it in a
smarter and cleaner way.